Sun, Oct 29, 2023, 2:59 PM +03:00

+++ credit/credit.c
@@ -15,7 +15,7 @@ int main(void)
     {
         if (power % 2 == 0)
         {
-            int x = n % pow(10 , power) - n % pow(10 , power-1);
+            long x = n % pow(10 , power) - n % pow(10 , power-1);
 
             if (10 < x) & (x > 100)
             {
